Break map initialization and rendering function out of map.js
